+++ src/components/RigidBodyComponents.cpp
@@ -39,8 +39,8 @@ RigidBodyComponents::RigidBodyComponents(MemoryAllocator& allocator)
                                 sizeof(bool) + sizeof(bool) + sizeof(decimal) + sizeof(BodyType) +
                                 sizeof(Vector3) + sizeof(Vector3) + sizeof(Vector3) +
                                 sizeof(Vector3) + sizeof(decimal) + sizeof(decimal) +
-                                sizeof(decimal) + sizeof(decimal) + sizeof(Matrix3x3) +
-                                sizeof(Matrix3x3) + sizeof(Vector3) + sizeof(Vector3) +
+                                sizeof(decimal) + sizeof(decimal) + sizeof(Vector3) +
+                                sizeof(Vector3) + sizeof(Vector3) + sizeof(Vector3) +
                                 sizeof(Vector3) + sizeof(Vector3) + sizeof(Vector3) +
                                 sizeof(Quaternion) + sizeof(Vector3) + sizeof(Vector3) +
                                 sizeof(bool) + sizeof(bool) + sizeof(List<Entity>)) {
@@ -76,8 +76,8 @@ void RigidBodyComponents::allocate(uint32 nbComponentsToAllocate) {
     decimal* newAngularDampings = reinterpret_cast<decimal*>(newLinearDampings + nbComponentsToAllocate);
     decimal* newMasses = reinterpret_cast<decimal*>(newAngularDampings + nbComponentsToAllocate);
     decimal* newInverseMasses = reinterpret_cast<decimal*>(newMasses + nbComponentsToAllocate);
-    Matrix3x3* newInertiaTensorLocal = reinterpret_cast<Matrix3x3*>(newInverseMasses + nbComponentsToAllocate);
-    Matrix3x3* newInertiaTensorLocalInverses = reinterpret_cast<Matrix3x3*>(newInertiaTensorLocal + nbComponentsToAllocate);
+    Vector3* newInertiaTensorLocal = reinterpret_cast<Vector3*>(newInverseMasses + nbComponentsToAllocate);
+    Vector3* newInertiaTensorLocalInverses = reinterpret_cast<Vector3*>(newInertiaTensorLocal + nbComponentsToAllocate);
     Vector3* newConstrainedLinearVelocities = reinterpret_cast<Vector3*>(newInertiaTensorLocalInverses + nbComponentsToAllocate);
     Vector3* newConstrainedAngularVelocities = reinterpret_cast<Vector3*>(newConstrainedLinearVelocities + nbComponentsToAllocate);
     Vector3* newSplitLinearVelocities = reinterpret_cast<Vector3*>(newConstrainedAngularVelocities + nbComponentsToAllocate);
@@ -108,8 +108,8 @@ void RigidBodyComponents::allocate(uint32 nbComponentsToAllocate) {
         memcpy(newAngularDampings, mAngularDampings, mNbComponents * sizeof(decimal));
         memcpy(newMasses, mMasses, mNbComponents * sizeof(decimal));
         memcpy(newInverseMasses, mInverseMasses, mNbComponents * sizeof(decimal));
-        memcpy(newInertiaTensorLocal, mLocalInertiaTensors, mNbComponents * sizeof(Matrix3x3));
-        memcpy(newInertiaTensorLocalInverses, mInverseInertiaTensorsLocal, mNbComponents * sizeof(Matrix3x3));
+        memcpy(newInertiaTensorLocal, mLocalInertiaTensors, mNbComponents * sizeof(Vector3));
+        memcpy(newInertiaTensorLocalInverses, mInverseInertiaTensorsLocal, mNbComponents * sizeof(Vector3));
         memcpy(newConstrainedLinearVelocities, mConstrainedLinearVelocities, mNbComponents * sizeof(Vector3));
         memcpy(newConstrainedAngularVelocities, mConstrainedAngularVelocities, mNbComponents * sizeof(Vector3));
         memcpy(newSplitLinearVelocities, mSplitLinearVelocities, mNbComponents * sizeof(Vector3));
@@ -178,8 +178,8 @@ void RigidBodyComponents::addComponent(Entity bodyEntity, bool isSleeping, const
     mAngularDampings[index] = decimal(0.0);
     mMasses[index] = decimal(1.0);
     mInverseMasses[index] = decimal(1.0);
-    new (mLocalInertiaTensors + index) Matrix3x3(1.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 1.0);
-    new (mInverseInertiaTensorsLocal + index) Matrix3x3(1.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 1.0);
+    new (mLocalInertiaTensors + index) Vector3(1.0, 1.0, 1.0);
+    new (mInverseInertiaTensorsLocal + index) Vector3(1.0, 1.0, 1.0);
     new (mConstrainedLinearVelocities + index) Vector3(0, 0, 0);
     new (mConstrainedAngularVelocities + index) Vector3(0, 0, 0);
     new (mSplitLinearVelocities + index) Vector3(0, 0, 0);
@@ -222,8 +222,8 @@ void RigidBodyComponents::moveComponentToIndex(uint32 srcIndex, uint32 destIndex
     mAngularDampings[destIndex] = mAngularDampings[srcIndex];
     mMasses[destIndex] = mMasses[srcIndex];
     mInverseMasses[destIndex] = mInverseMasses[srcIndex];
-    new (mLocalInertiaTensors + destIndex) Matrix3x3(mLocalInertiaTensors[srcIndex]);
-    new (mInverseInertiaTensorsLocal + destIndex) Matrix3x3(mInverseInertiaTensorsLocal[srcIndex]);
+    new (mLocalInertiaTensors + destIndex) Vector3(mLocalInertiaTensors[srcIndex]);
+    new (mInverseInertiaTensorsLocal + destIndex) Vector3(mInverseInertiaTensorsLocal[srcIndex]);
     new (mConstrainedLinearVelocities + destIndex) Vector3(mConstrainedLinearVelocities[srcIndex]);
     new (mConstrainedAngularVelocities + destIndex) Vector3(mConstrainedAngularVelocities[srcIndex]);
     new (mSplitLinearVelocities + destIndex) Vector3(mSplitLinearVelocities[srcIndex]);
@@ -265,8 +265,8 @@ void RigidBodyComponents::swapComponents(uint32 index1, uint32 index2) {
     decimal angularDamping1 = mAngularDampings[index1];
     decimal mass1 = mMasses[index1];
     decimal inverseMass1 = mInverseMasses[index1];
-    Matrix3x3 inertiaTensorLocal1 = mLocalInertiaTensors[index1];
-    Matrix3x3 inertiaTensorLocalInverse1 = mInverseInertiaTensorsLocal[index1];
+    Vector3 inertiaTensorLocal1 = mLocalInertiaTensors[index1];
+    Vector3 inertiaTensorLocalInverse1 = mInverseInertiaTensorsLocal[index1];
     Vector3 constrainedLinearVelocity1(mConstrainedLinearVelocities[index1]);
     Vector3 constrainedAngularVelocity1(mConstrainedAngularVelocities[index1]);
     Vector3 splitLinearVelocity1(mSplitLinearVelocities[index1]);
@@ -336,8 +336,8 @@ void RigidBodyComponents::destroyComponent(uint32 index) {
     mAngularVelocities[index].~Vector3();
     mExternalForces[index].~Vector3();
     mExternalTorques[index].~Vector3();
-    mLocalInertiaTensors[index].~Matrix3x3();
-    mInverseInertiaTensorsLocal[index].~Matrix3x3();
+    mLocalInertiaTensors[index].~Vector3();
+    mInverseInertiaTensorsLocal[index].~Vector3();
     mConstrainedLinearVelocities[index].~Vector3();
     mConstrainedAngularVelocities[index].~Vector3();
     mSplitLinearVelocities[index].~Vector3();